requirements.txt
openai
requests
PyMuPDF
python-dotenv
tiktoken
Flask
flask-cors